requirements.txt
rapidocr-onnxruntime
openpyxl>=3.1.0
pillow>=10.0.0
pyautogui>=0.9.54
opencv-python>=4.6.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
pandas>=1.3.0
jieba>=0.42.1
